+++ AI-Social-Platform.Server/Controllers/PublicationController.cs
@@ -108,191 +108,4 @@ public class PublicationController : ControllerBase
             return BadRequest(ex.Message);
         }
     }
-
-    [HttpGet("comment/{publicationId}", Name = "getComment")]
-    public async Task<IEnumerable<CommentDto>> GetComments(Guid publicationId)
-    {
-        try
-        {
-            var comments = await publicationService.GetCommentsOnPublicationAsync(publicationId);
-            return comments;
-        }
-        catch (Exception ex)
-        {
-            throw new Exception(ex.Message);
-        }
-    }
-
-    [HttpPost("comment", Name = "createComment")]
-    public async Task<IActionResult> CreateComment(CommentFormDto dto)
-    {
-        try
-        {
-            await publicationService.CreateCommentAsync(dto);
-            return CreatedAtAction(nameof(CreateComment), new { message = CommentSuccessfullyCreated});
-        }
-        catch (NullReferenceException ex)
-        {
-            return NotFound(ex.Message);
-        }
-        catch (Exception ex)
-        {
-            return BadRequest(ex.Message);
-        }
-    }
-
-    [HttpDelete("comment/{commentId}", Name = "deleteComment")]
-    public async Task<IActionResult> DeleteComment(Guid commentId)
-    {
-        try
-        {
-            await publicationService.DeleteCommentAsync(commentId);
-            return Ok(CommentSuccessfullyDeleted);
-        }
-        catch (NullReferenceException ex)
-        {
-            return NotFound(ex.Message);
-        }
-        catch (AccessViolationException ex)
-        {
-            return StatusCode(403, ex.Message);
-        }
-        catch (Exception ex)
-        {
-            return BadRequest(ex.Message);
-        }
-    }
-
-
-    [HttpPut("comment/{commentId}", Name = "updateComment")]
-    public async Task<IActionResult> UpdateComment(CommentFormDto dto, Guid commentId)
-    {
-        try
-        {
-            await publicationService.UpdateCommentAsync(dto, commentId);
-            return Ok(CommentSuccessfullyEdited);
-        }
-        catch (NullReferenceException ex)
-        {
-            return NotFound(ex.Message);
-        }
-        catch (AccessViolationException ex)
-        {
-            return StatusCode(403, ex.Message);
-        }
-        catch (Exception ex)
-        {
-            return BadRequest(ex.Message);
-        }
-    }
-
-    //Likes
-    [HttpGet("like/{publicationId}", Name = "getLikes")]
-    public async Task<IEnumerable<LikeDto>> GetLikes(Guid publicationId)
-    {
-        try
-        {
-            var likes = await publicationService.GetLikesOnPublicationAsync(publicationId);
-            return likes;
-        }
-        catch (Exception ex)
-        {
-            throw new Exception(ex.Message);
-        }
-    }
-
-    [HttpPost("like/{publicationId}", Name = "createLike")]
-    public async Task<IActionResult> CreateLike(Guid publicationId)
-    {
-        try
-        {
-            await publicationService.CreateLikesOnPublicationAsync(publicationId);
-            return CreatedAtAction(nameof(CreateLike), new { message = LikeSuccessfullyCreated });
-        }
-        catch (NullReferenceException ex)
-        {
-            return NotFound(ex.Message);
-        }
-        catch (Exception ex)
-        {
-            return BadRequest(ex.Message);
-        }
-    }
-
-    [HttpDelete("like/{likeId}", Name = "deleteLike")]
-    public async Task<IActionResult> DeleteLike(Guid likeId)
-    {
-        try
-        {
-            await publicationService.DeleteLikeOnPublicationAsync(likeId);
-            return Ok(LikeSuccessfullyDeletedFromPublication);
-        }
-        catch (NullReferenceException ex)
-        {
-            return NotFound(ex.Message);
-        }
-        catch (AccessViolationException ex)
-        {
-            return StatusCode(403, ex.Message);
-        }
-        catch (Exception ex)
-        {
-            return BadRequest(ex.Message);
-        }
-    }
-
-    //Shares
-    [HttpGet("share/{publicationId}", Name = "getShares")]
-    public async Task<IEnumerable<ShareDto>> GetShares(Guid publicationId)
-    {
-        try
-        {
-            var shares = await publicationService.GetSharesOnPublicationAsync(publicationId);
-            return shares;
-        }
-        catch (Exception ex)
-        {
-            throw new Exception(ex.Message);
-        }
-    }
-
-    [HttpPost("share/{publicationId}", Name = "createShare")]
-    public async Task<IActionResult> CreateShare(Guid publicationId)
-    {
-        try
-        {
-            await publicationService.CreateSharesOnPublicationAsync(publicationId);
-            return CreatedAtAction(nameof(CreateShare), new { message = ShareSuccessfullyCreated });
-        }
-        catch (NullReferenceException ex)
-        {
-            return NotFound(ex.Message);
-        }
-        catch (Exception ex)
-        {
-            return BadRequest(ex.Message);
-        }
-    }
-
-    [HttpDelete("share/{shareId}", Name = "deleteShare")]
-    public async Task<IActionResult> DeleteShare(Guid shareId)
-    {
-        try
-        {
-            await publicationService.DeleteShareOnPublicationAsync(shareId);
-            return Ok(ShareSuccessfullyDeletedFromPublication);
-        }
-        catch (NullReferenceException ex)
-        {
-            return NotFound(ex.Message);
-        }
-        catch (AccessViolationException ex)
-        {
-            return StatusCode(403, ex.Message);
-        }
-        catch (Exception ex)
-        {
-            return BadRequest(ex.Message);
-        }
-    }
 }
