+++ AI-Social-Platform.Server/Controllers/SocialFeatureController.cs
@@ -1,7 +1,11 @@
-﻿using AI_Social_Platform.Services.Data.Interfaces;
+﻿using AI_Social_Platform.Services.Data;
+using AI_Social_Platform.Services.Data.Interfaces;
+using AI_Social_Platform.Services.Data.Models.PublicationDtos;
 using AI_Social_Platform.Services.Data.Models.SocialFeature;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+
+using static AI_Social_Platform.Common.NotificationMessagesConstants;
 
 namespace AI_Social_Platform.Server.Controllers
 {
@@ -42,6 +46,193 @@ namespace AI_Social_Platform.Server.Controllers
             catch (Exception ex)
             {
                 throw new Exception(ex.Message);
+            }
+        }
+
+        [HttpGet("comment/{publicationId}", Name = "getComment")]
+        public async Task<IEnumerable<CommentDto>> GetComments(Guid publicationId)
+        {
+            try
+            {
+                var comments = await baseSocialService.GetCommentsOnPublicationAsync(publicationId);
+                return comments;
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
+        }
+
+        [HttpPost("comment", Name = "createComment")]
+        public async Task<IActionResult> CreateComment(CommentFormDto dto)
+        {
+            try
+            {
+                await baseSocialService.CreateCommentAsync(dto);
+                return CreatedAtAction(nameof(CreateComment), new { Message = CommentSuccessfullyCreated });
+            }
+            catch (NullReferenceException ex)
+            {
+                return NotFound(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
+        }
+
+        [HttpDelete("comment/{commentId}", Name = "deleteComment")]
+        public async Task<IActionResult> DeleteComment(Guid commentId)
+        {
+            try
+            {
+                await baseSocialService.DeleteCommentAsync(commentId);
+                return Ok(new {Message = CommentSuccessfullyDeleted });
+            }
+            catch (NullReferenceException ex)
+            {
+                return NotFound(ex.Message);
+            }
+            catch (AccessViolationException ex)
+            {
+                return StatusCode(403, ex.Message);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
+        }
+
+
+        [HttpPut("comment/{commentId}", Name = "updateComment")]
+        public async Task<IActionResult> UpdateComment(CommentFormDto dto, Guid commentId)
+        {
+            try
+            {
+                await baseSocialService.UpdateCommentAsync(dto, commentId);
+                return Ok(new {Message = CommentSuccessfullyEdited });
+            }
+            catch (NullReferenceException ex)
+            {
+                return NotFound(ex.Message);
+            }
+            catch (AccessViolationException ex)
+            {
+                return StatusCode(403, ex.Message);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
+        }
+
+        //Likes
+        [HttpGet("like/{publicationId}", Name = "getLikes")]
+        public async Task<IEnumerable<LikeDto>> GetLikes(Guid publicationId)
+        {
+            try
+            {
+                var likes = await baseSocialService.GetLikesOnPublicationAsync(publicationId);
+                return likes;
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
+        }
+
+        [HttpPost("like/{publicationId}", Name = "createLike")]
+        public async Task<IActionResult> CreateLike(Guid publicationId)
+        {
+            try
+            {
+                await baseSocialService.CreateLikesOnPublicationAsync(publicationId);
+                return CreatedAtAction(nameof(CreateLike), new { Message = LikeSuccessfullyCreated });
+            }
+            catch (NullReferenceException ex)
+            {
+                return NotFound(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
+        }
+
+        [HttpDelete("like/{likeId}", Name = "deleteLike")]
+        public async Task<IActionResult> DeleteLike(Guid likeId)
+        {
+            try
+            {
+                await baseSocialService.DeleteLikeOnPublicationAsync(likeId);
+                return Ok(new {Message = LikeSuccessfullyDeletedFromPublication });
+            }
+            catch (NullReferenceException ex)
+            {
+                return NotFound(ex.Message);
+            }
+            catch (AccessViolationException ex)
+            {
+                return StatusCode(403, ex.Message);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
+        }
+
+        //Shares
+        [HttpGet("share/{publicationId}", Name = "getShares")]
+        public async Task<IEnumerable<ShareDto>> GetShares(Guid publicationId)
+        {
+            try
+            {
+                var shares = await baseSocialService.GetSharesOnPublicationAsync(publicationId);
+                return shares;
+            }
+            catch (Exception ex)
+            {
+                throw new Exception(ex.Message);
+            }
+        }
+
+        [HttpPost("share/{publicationId}", Name = "createShare")]
+        public async Task<IActionResult> CreateShare(Guid publicationId)
+        {
+            try
+            {
+                await baseSocialService.CreateSharesOnPublicationAsync(publicationId);
+                return CreatedAtAction(nameof(CreateShare), new { Message = ShareSuccessfullyCreated });
+            }
+            catch (NullReferenceException ex)
+            {
+                return NotFound(ex.Message);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
+            }
+        }
+
+        [HttpDelete("share/{shareId}", Name = "deleteShare")]
+        public async Task<IActionResult> DeleteShare(Guid shareId)
+        {
+            try
+            {
+                await baseSocialService.DeleteShareOnPublicationAsync(shareId);
+                return Ok(new {Message = ShareSuccessfullyDeletedFromPublication });
+            }
+            catch (NullReferenceException ex)
+            {
+                return NotFound(ex.Message);
+            }
+            catch (AccessViolationException ex)
+            {
+                return StatusCode(403, ex.Message);
+            }
+            catch (Exception ex)
+            {
+                return BadRequest(ex.Message);
             }
         }
