+++ AI-Social-Platform.Services.Data/Interfaces/IPublicationService.cs
@@ -11,30 +11,5 @@ public interface IPublicationService
     public Task UpdatePublicationAsync(PublicationFormDto dto, Guid id);
 
     public Task DeletePublicationAsync(Guid id);
-
-    //Comments
-    public Task<IEnumerable<CommentDto>> GetCommentsOnPublicationAsync(Guid publicationId);
-
-    public Task CreateCommentAsync(CommentFormDto dto);
-
-    public Task UpdateCommentAsync(CommentFormDto dto, Guid id);
-
-    public Task DeleteCommentAsync(Guid id);
-
-    //Likes
-    public Task<IEnumerable<LikeDto>> GetLikesOnPublicationAsync(Guid publicationId);
-    public Task CreateLikesOnPublicationAsync(Guid publicationId);
-
-    public Task DeleteLikeOnPublicationAsync(Guid likeId);
-
-    //Shares
-    public Task<IEnumerable<ShareDto>> GetSharesOnPublicationAsync(Guid publicationId);
-    public Task CreateSharesOnPublicationAsync(Guid publicationId);
-
-    public Task DeleteShareOnPublicationAsync(Guid shareId);
-
-
-
-  
 }
